chore: rename socket-server to event-server
